fix(qa): split oversized scenario catalog tests
